+++ src/requirements.txt
@@ -5,8 +5,8 @@ colorama==0.4.6
 css-inline==0.8.7
 dnspython==1.16.0
 eventlet==0.30.2
-face-recognition==1.3.0
-face-recognition-models==0.3.0
+# face-recognition==1.3.0
+# face-recognition-models==0.3.0
 Flask==2.0.3
 Flask-Admin==1.6.1
 Flask-Login==0.6.2
