+++ src/views/index.py
@@ -9,7 +9,10 @@ from flask import Blueprint, redirect, url_for
 from flask_admin import expose
 from flask_login import logout_user
 
-from models import MyAdminIndexView
+import sys
+sys.path.append("..")
+
+from configurations.models import MyAdminIndexView
 
 index = Blueprint("index", __name__, template_folder="./templates")
