+++ src/views/login.py
@@ -11,11 +11,13 @@ If the user is not authorized or enters incorrect login credentials, an danger m
 from flask import Blueprint, flash, redirect, render_template, request, url_for
 from flask_login import LoginManager, login_user
 
-from models import Admins, Lecturers, Students
+import sys
+sys.path.append("..")
+
+from configurations.models import Admins, Lecturers, Students
 
 login = Blueprint(
-    "login", __name__, template_folder="./templates", static_folder="./templates"
-)
+    "login", __name__)
 
 login_manager = LoginManager()
 login_manager.init_app(login)
